+++ src/ts/Ads/Native/WebPlayer.ts
@@ -43,7 +43,7 @@ export enum WebplayerEvent {
     FORM_RESUBMISSION,
     UNHANDLED_KEY_EVENT,
     SHOULD_INTERCEPT_REQUEST,
-    CREATE_WEBVIEW,
+    CREATE_WEBVIEW, // Only exists on iOS
     FRAME_UPDATE,
     GET_FRAME_RESPONSE
 }
@@ -259,6 +259,7 @@ export class WebPlayerApi extends NativeApi {
         });
     }
 
+    // parameters.pop is used here because the params length is variable and we want the last param
     public handleEvent(event: string, parameters: unknown[]): void {
         switch (event) {
             case WebplayerEvent[WebplayerEvent.PAGE_STARTED]: