+++ src/ts/Ads/Views/OpenMeasurement/OpenMeasurement.ts
@@ -269,6 +269,7 @@ export class OpenMeasurement<T extends Campaign> extends View<T> {
             // TODO: Refactor. Admob Code Path
             this._sessionStartEventData = sessionEvent;
             this._sessionStartEventData.data.vendorkey = this._vendorKey;
+            this._sessionStartEventData.data.verificationParameters = this._verificationVendorMap[this._vendorKey];
             this._omBridge.triggerSessionEvent(this._sessionStartEventData);
         }
     }